fix new category update
