+++ Phrasebook/CategoryGridView.swift
@@ -1,25 +1,25 @@
 //
-//  CategoryGridView.swift
+//  ContentView.swift
 //  Phrasebook
 //
-//  Created by Christine Røde on 05/09/2024.
+//  Created by Christine Røde on 08/09/2024.
 //
 
-import Foundation
 import SwiftUI
 
 struct CategoryGridView: View {
-    let categories: FetchedResults<Category>
+    let categories: [Category]
     let colors: [Color]
     let editCategory: (Category) -> Void
     let deleteCategory: (Category) -> Void
     let presentNewCategorySheet: () -> Void
+    let currentLanguage: Language?
 
     var body: some View {
         LazyVGrid(columns: [GridItem(.adaptive(minimum: 150))], spacing: 16) {
             ForEach(categories, id: \.self) { category in
                 PressableCategoryTile(
-                    destination: PhrasesView(category: category, currentLanguage: LanguageManager.shared.currentLanguage),
+                    destination: PhrasesView(category: category, currentLanguage: currentLanguage),
                     label: (category.name ?? "Unknown Category"),
                     symbol: Image(systemName: category.symbol ?? "folder")
                 )
@@ -55,24 +55,6 @@ struct CategoryGridView: View {
                 .cornerRadius(16)
             }
             .zIndex(100)
-        }
-    }
-}
-
-struct EmptyStateView: View {
-    var presentNewCategorySheet: () -> Void  // Closure to present the new category sheet
-    
-    var body: some View {
-        VStack {
-            Text("No Categories")
-                .font(.title2)
-                .foregroundStyle(.secondary)
-            Button("Add New", systemImage: "plus", action: presentNewCategorySheet)
-                .padding()
-                .buttonStyle(.borderedProminent)
-                .controlSize(.regular)
-                .bold()
-                .cornerRadius(50)
         }
     }
 }
